backend/requirements.txt
# Django Core
Django==4.2.9
djangorestframework==3.14.0

# MongoDB Integration (using compatible versions)
pymongo==4.6.1
dnspython==2.5.0

# Environment Variables
python-dotenv==1.0.0

# CORS Headers
django-cors-headers==4.3.1

# Security & Production
gunicorn==21.2.0
whitenoise==6.6.0
